feat(transform): add shell transform
