+++ src/module/TransformModule.ts
@@ -2,6 +2,7 @@ import { ModuleOptions } from 'noicejs';
 
 import { FlattenTransform } from '../transform/FlattenTransform';
 import { JsonpathTransform } from '../transform/JsonpathTransform';
+import { ShellTransform } from '../transform/ShellTransform';
 import { TemplateTransform } from '../transform/TemplateTransform';
 import { BaseModule } from './BaseModule';
 
@@ -11,6 +12,7 @@ export class TransformModule extends BaseModule {
 
     this.bindService(FlattenTransform);
     this.bindService(JsonpathTransform);
+    this.bindService(ShellTransform);
     this.bindService(TemplateTransform);
   }
 }
